- [x] Rely less on CardanoApi in prop_balanceTransactionValid
- [x] Generate fewer PParam updates in tests
- [x] Use cheaper PParam gen for the sake of hashing

### Comments


`cabal test cardano-wallet:unit --test-options='-j 8 --match "valid transactions or fails"'`

#### This PR

```
Cardano.Wallet.Shelley.Transaction
  balanceTransaction
    produces valid transactions or fails [✔] (11278ms)
      +++ OK, passed 1000 tests:
      100.0% has payment outputs
      55.0% >5 payment outputs
      33.8% >10 payment outputs
       9.6% >20 payment outputs
       9.1% partial tx had zero ada outputs

      49.1% existing total collateral
      26.1% existing collateral return outputs
      10.1% output below minCoinValue
       5.6% missing tokens
       3.9% conflicting networks
       3.1% success
       0.7% missing coin
       0.5% existing collateral
       0.4% maxTxSize limit exceeded
       0.3% missing coin and tokens
       0.1% ReferenceScriptsNotSupported
       0.1% failed with ByronTxOutInContext

Finished in 11.2793 seconds, used 12.2644 seconds of CPU time
1 example, 0 failures
```

#### Master (376280e420)

```
Cardano.Wallet.Shelley.Transaction
  balanceTransaction
    produces valid transactions or fails [✔] (77431ms)
      +++ OK, passed 1000 tests:
      100.0% has payment outputs
      55.8% >5 payment outputs
      35.3% >10 payment outputs
      10.1% partial tx had zero ada outputs
       9.3% >20 payment outputs

      50.9% existing total collateral
      24.9% existing collateral return outputs
      10.8% output below minCoinValue
       4.7% missing tokens
       4.0% conflicting networks
       2.7% success
       1.0% missing coin
       0.5% missing coin and tokens
       0.2% existing collateral
       0.1% failed with ByronTxOutInContext
       0.1% maxTxSize limit exceeded
       0.1% unable to construct change

Finished in 77.4322 seconds, used 84.2236 seconds of CPU time
1 example, 0 failures
```

### Issue Number

<!-- Reference the Jira/GitHub issue that this PR relates to, and which requirements it tackles.
  Note: Jira issues of the form ADP- will be auto-linked. -->