@jonathanknowles 2023-05-16
Hi @Anviking 

I noticed that `master` has coverage of "unable to construct change", but with this PR, that case seems not to be covered? (According to your PR description.)