@Anviking 2023-05-16
> unable to construct change

Good catch. While I can imagine this PR reducing the coverage for `unable to construct change`, I am unable reproducing it again on `master`. I.e. the coverage of this label is already very very low.

I'd look into this some time after the balanceTx library, closer to being able to test `selectAssets` and rewriting `Cardano.Api.Gen` generators with ledger ones for more performance and maxSuccess.